adding exception methods and updating swagger.yaml

+++ src/services/foodService.js
@@ -1,4 +1,5 @@
 const { FoodRepo } = require('../repositories/db');
+const HttpError = require('../errors/HttpError');
 
 exports.createFood = async (data) => {
   const food = await FoodRepo.create(data);
@@ -10,9 +11,13 @@ exports.getFoods = async () => {
 };
 
 exports.updateFood = async (id, data) => {
-  return FoodRepo.update(id, data);
+  const updated = await FoodRepo.update(id, data);
+  if (!updated) throw new HttpError(404, 'Alimento não encontrado');
+  return updated;
 };
 
 exports.deleteFood = async (id) => {
+  const existed = await FoodRepo.findById(id);
+  if (!existed) throw new HttpError(404, 'Alimento não encontrado');
   await FoodRepo.delete(id);
 };
